+++ test/fused/index_of_test.cpp
@@ -15,15 +15,15 @@ TEST(index_of, canGetIndexOfSimpleType)
   char a = '1';
   int b = 2;
   float c = 3;
-  EXPECT_EQ(0, type_index<char>(a, b, c, b));
-  EXPECT_EQ(1, type_index<int>(a, b, c, b));
-  EXPECT_EQ(2, type_index<float>(a, b, c, b));
+  EXPECT_EQ(0u, static_cast<std::size_t>(type_index<char>(a, b, c, b)));
+  EXPECT_EQ(1u, static_cast<std::size_t>(type_index<int>(a, b, c, b)));
+  EXPECT_EQ(2u, static_cast<std::size_t>(type_index<float>(a, b, c, b)));
 }
 
 TEST(index_of, canUseIndexOfWithTypeLists)
 {
-  EXPECT_EQ(0, index_of<char>(type_list<char, int, float>));
-  EXPECT_EQ(1, index_of<int>(type_list<char, int, float>));
+  EXPECT_EQ(0u, static_cast<std::size_t>(index_of<char>(type_list<char, int, float>)));
+  EXPECT_EQ(1u, static_cast<std::size_t>(index_of<int>(type_list<char, int, float>)));
   static_assert(0 == index_of<char>(type_list<char, int, float, int, double>));
   static_assert(1 == index_of<int>(type_list<char, int, float, int, double>));
   static_assert(2 == index_of<float>(type_list<char, int, float, int, double>));
@@ -34,7 +34,7 @@ TEST(index_of, canUseIndexOfWithTypeLists)
 TEST(index_of, canUseIndexOfInConsexpr)
 {
   constexpr auto idx = index_of<char>(type_list<char, int, float>);
-  EXPECT_EQ(0, idx);
+  EXPECT_EQ(0u, static_cast<std::size_t>(idx));
   constexpr auto cnt = index_of<int>(type_list<char, int, float>);
   std::array<int, cnt> arr;
   EXPECT_EQ(sizeof(int), sizeof(arr));
